Fix accounts-related imports
