+++ project/accounts/api.py
@@ -2,9 +2,10 @@ from django.apps import AppConfig
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.decorators import action
 from rest_framework.response import Response
-from ..permissions import IsAccountOwnerOrDuringRegistrationOrReadOnly
+from api.permissions import IsAccountOwnerOrDuringRegistrationOrReadOnly
 
-from api.models import Thread, Account
+from api.models import Thread
+from accounts.models import Account
 from api.serializers import (
     ThreadSerializer,
     CategorySerializer,
