+++ project/accounts/models.py
@@ -240,3 +240,17 @@ class Account(models.Model):
             return True
         else:
             return False
+
+@deconstructible
+class PathAndRename(object):
+    def __init__(self, sub_path):
+        self.sub_path = sub_path
+
+    def __call__(self, instance, filename):
+        extension = filename.split(".")[-1]
+        new_filename = str(uuid.uuid4())
+        filename = "{}.{}".format(new_filename, extension)
+        return os.path.join(self.sub_path, filename)
+
+
+profile_upload_path = PathAndRename("")
