+++ project/api/forms.py
@@ -1,7 +1,7 @@
 from django import forms
 from django.core.files.images import get_image_dimensions
 from django.contrib.auth import get_user_model
-from .models import Account
+from accounts.models import Account
 
 User = get_user_model()
