+++ project/api/models/activity.py
@@ -5,7 +5,7 @@ Records user activity
 
 from django.db import models
 
-from .account import Account
+from accounts.models import Account
 from .civi import Civi
 from .thread import Thread
