+++ project/api/models/response.py
@@ -5,7 +5,7 @@ Records user response
 
 from django.db import models
 
-from .account import Account
+from accounts.models import Account
 from .civi import Civi
