+++ project/api/write.py
@@ -25,7 +25,8 @@ from django.contrib.sites.shortcuts import get_current_site
 from api.forms import UpdateProfileImage
 from api.models import Thread
 from accounts.utils import send_mass_email
-from .models import Account, Activity, Category, Civi, CiviImage
+from .models import Activity, Category, Civi, CiviImage
+from accounts.models import Account
 from core.custom_decorators import require_post_params
 from core.constants import US_STATES
